+++ src/pland/land/observer/ILandObserver.h
@@ -1,0 +1,28 @@
+#pragma once
+#include <memory>
+
+namespace mce {
+class UUID;
+}
+
+namespace land {
+class Land;
+}
+
+namespace land::observer {
+
+class ILandObserver {
+public:
+    virtual ~ILandObserver() = default;
+
+    virtual void
+    onOwnerChanged(std::shared_ptr<Land> const& land, mce::UUID const& oldOwner, mce::UUID const& newOwner) = 0;
+
+    virtual void onMemberAdded(std::shared_ptr<Land> const& land, mce::UUID const& member) = 0;
+
+    virtual void onMemberRemoved(std::shared_ptr<Land> const& land, mce::UUID const& member) = 0;
+
+    virtual void onMembersCleared(std::shared_ptr<Land> const& land) = 0;
+};
+
+} // namespace land::observer